# scripts/openapi-snapshot.json
{
 "components": {
  "schemas": {
...
   "Body_upload_report_api_v1_reports_upload_post": {
    "properties": {
     "file": {
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "type": "string"
     },
     "programme": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Programme"
     },
     "sender": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender"
     }
    },
    "required": [
     "file"
    ],
    "title": "Body_upload_report_api_v1_reports_upload_post",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobOut": {
    "properties": {
     "attempts": {
      "title": "Attempts",
      "type": "integer"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "finished_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "heartbeat_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Heartbeat At"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "max_attempts": {
      "title": "Max Attempts",
      "type": "integer"
     },
     "metrics": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Metrics"
     },
     "stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stage"
     },
     "started_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "id",
     "kind",
     "status",
     "stage",
     "attempts",
     "max_attempts",
     "error_code",
     "error_message",
     "created_at",
     "started_at",
     "finished_at",
     "heartbeat_at"
    ],
    "title": "JobOut",
    "type": "object"
   },
...
   "Page_ReportOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ReportOut"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     },
     "offset": {
      "title": "Offset",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[ReportOut]",
    "type": "object"
   },
   "ReportOut": {
    "properties": {
     "coverage_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Coverage Pct"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "expires_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     },
     "file_kind": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "File Kind"
     },
     "file_name": {
      "title": "File Name",
      "type": "string"
     },
     "file_size_bytes": {
      "title": "File Size Bytes",
      "type": "integer"
     },
     "grade": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "ingest_notes": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingest Notes"
     },
     "issues_found": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Issues Found"
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "processing_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processing Error"
     },
     "programme": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Programme"
     },
     "rows_processed": {
      "title": "Rows Processed",
      "type": "integer"
     },
     "rows_total": {
      "title": "Rows Total",
      "type": "integer"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "sender": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender"
     },
     "sheet_count_total": {
      "title": "Sheet Count Total",
      "type": "integer"
     },
     "source_channel": {
      "default": "upload",
      "title": "Source Channel",
      "type": "string"
     },
     "source_sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Sha256"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "updated_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     }
    },
    "required": [
     "id",
     "created_at",
     "file_name",
     "file_size_bytes",
     "sheet_count_total",
     "rows_processed",
     "rows_total",
     "coverage_pct",
     "grade",
     "score",
     "status"
    ],
    "title": "ReportOut",
    "type": "object"
   },
   "ReportSummaryOut": {
    "properties": {
     "report": {
      "$ref": "#/components/schemas/ReportOut"
     },
     "summary": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     }
    },
    "required": [
     "report",
     "summary"
    ],
    "title": "ReportSummaryOut",
    "type": "object"
   },
...
   "TemplateCreateRequest": {
    "additionalProperties": false,
    "properties": {
     "field_mappings": {
      "additionalProperties": {
       "type": "string"
      },
      "maxProperties": 200,
      "title": "Field Mappings",
      "type": "object"
     },
     "name": {
      "maxLength": 255,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "sender_identifier": {
      "anyOf": [
       {
        "maxLength": 255,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender Identifier"
     }
    },
    "required": [
     "name",
     "field_mappings"
    ],
    "title": "TemplateCreateRequest",
    "type": "object"
   },
   "TemplateOut": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     },
     "field_mappings": {
      "additionalProperties": true,
      "title": "Field Mappings",
      "type": "object"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "is_deletable": {
      "title": "Is Deletable",
      "type": "boolean"
     },
     "is_standard": {
      "title": "Is Standard",
      "type": "boolean"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "sender_identifier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sender Identifier"
     }
    },
    "required": [
     "id",
     "name",
     "sender_identifier",
     "field_mappings",
     "is_standard",
     "is_deletable",
     "created_by",
     "created_at"
    ],
    "title": "TemplateOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "TrueBind API",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/reports": {
   "get": {
    "operationId": "list_reports_api_v1_reports_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "maximum": 10000000,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ReportOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Reports",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v1/reports/upload": {
   "post": {
    "operationId": "upload_report_api_v1_reports_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_report_api_v1_reports_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Upload Report",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v1/reports/{report_id}": {
   "delete": {
    "operationId": "delete_report_api_v1_reports__report_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Report",
    "tags": [
     "reports"
    ]
   },
   "get": {
    "operationId": "get_report_api_v1_reports__report_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Report",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/cancel": {
   "post": {
    "operationId": "cancel_report_api_v1_reports__report_id__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cancel Report",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/export/audit.csv": {
   "get": {
    "operationId": "export_audit_api_v1_reports__report_id__export_audit_csv_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Audit",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v1/reports/{report_id}/export/claims.csv": {
   "get": {
    "description": "One line per extracted claim row with lineage (sheet + source row),\nunmapped source values and the findings raised against it.",
    "operationId": "export_claims_api_v1_reports__report_id__export_claims_csv_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Claims",
    "tags": [
     "reports"
    ]
   }
  },
  "/api/v1/reports/{report_id}/export/exceptions.csv": {
   "get": {
    "operationId": "export_exceptions_api_v1_reports__report_id__export_exceptions_csv_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Export Exceptions",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/retry": {
   "post": {
    "operationId": "retry_report_api_v1_reports__report_id__retry_post",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Retry Report",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/reports/{report_id}/summary": {
   "get": {
    "operationId": "get_report_summary_api_v1_reports__report_id__summary_get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportSummaryOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Report Summary",
    "tags": [
     "reports"
    ]
   }
  },
...
  "/api/v1/templates": {
   "get": {
    "operationId": "list_templates_api_v1_templates_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TemplateOut"
         },
         "title": "Response List Templates Api V1 Templates Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Templates",
    "tags": [
     "templates"
    ]
   },
   "post": {
    "operationId": "create_template_api_v1_templates_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TemplateCreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TemplateOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Template",
    "tags": [
     "templates"
    ]
   }
  },
  "/api/v1/templates/{template_id}": {
   "get": {
    "operationId": "get_template_api_v1_templates__template_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "template_id",
      "required": true,
      "schema": {
       "title": "Template Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TemplateOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Template",
    "tags": [
     "templates"
    ]
   }
  },
...
 }
}